feat: add shields.io badge spec
This adds a new shields.io badge:

- after merge: ![Pact Badge](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/pact-foundation/.github/master/assets/badge/v1.json)
- before merge: ![Pact Badge](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/jp-ellis/pact-foundation-dotgithub/feat/pact-badge/assets/badge/v1.json)

It can be used as follows:

```markdown
![Pact Badge](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/pact-foundation/.github/master/assets/badge/v1.json)
```

Unfortunately, there does not appear to be a way of overwriting the message part of the label (the right hand side), so I have created variations. I am assuming that if a version of the Pact spec is supported, all prior versions will be supported as well.

Signed-off-by: JP-Ellis <josh@jpellis.me>